requirements/requirements.txt
# Core components
apache-airflow-providers-apache-spark==4.1.5
apache-airflow-providers-openlineage==1.8.0
apache-airflow-providers-apache-kafka==1.1.0
apache-airflow-providers-amazon==7.3.0
kafka-python==2.0.2

# Data processing
pandas==1.5.3
numpy==1.24.2
pyarrow==12.0.0

# MinIO/S3 interaction
boto3==1.26.151
s3fs==2023.5.0

# Database clients
psycopg2-binary==2.9.6
pymongo==4.3.3
redis==4.5.5

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
requests==2.31.0
